+++ temp_cata/BetterBlizzFrames.lua
@@ -1365,7 +1365,7 @@ Frame:SetScript("OnEvent", function(...)
                 BBF.HideFrames()
                 --BBF.HookUnitFrameName()
             end)
-            if BetterBlizzFramesDB.partyCastbars then
+            if BetterBlizzFramesDB.partyCastbars or BetterBlizzFramesDB.petCastbar then
                 BBF.CreateCastbars()
             end